feat: parametrize multi serving
