+++ .env.serve.example
@@ -1,7 +1,8 @@
 HF_AUTH_TOKEN=
-LANGUAGE_CLASSIFIER_REPO_ID=Hexastack/language-classifier-cnn
-INTENT_CLASSIFIER_REPO_ID=Hexastack/intent-classifier-lstm
-SLOT_FILLER_REPO_ID=Hexastack/slot-filler-lstm
-MODEL_PATHS={"slot": "${SLOT_FILLER_REPO_ID}", "language": "${LANGUAGE_CLASSIFIER_REPO_ID}", "intent": "${INTENT_CLASSIFIER_REPO_ID}"}
+MODEL_PATHS='{
+    "language": "Hexastack/language-classifier-cnn", 
+    "intent": "Hexastack/intent-classifier-lstm"
+    }'
 PORT=5000
 HOST=0.0.0.0
+MODE=huggingface
