--- docker-compose.serve-local.yml
@@ -1,22 +1,0 @@
-version: '3.8'
-
-services:
-  ludwig:
-    image: ludwigai/ludwig:master
-    # You can use an environment variable instead ${MODEL_PATHS}, see docker-compose.serve-hf.yml file.
-    command: >-
-      multi_serve_local
-      --model_paths '{"language": "/results/experiment_run_2/model", "intent": "/results/experiment_run_1/model"}'  
-  
-    stdin_open: true
-    tty: true
-    env_file: .env
-    volumes:
-      - ./data:/data 
-      - ./results:/results
-      - ./cli.py:/usr/local/lib/python3.8/site-packages/ludwig/cli.py
-      - ./multi_serve_local.py:/usr/local/lib/python3.8/site-packages/ludwig/multi_serve_local.py
-
-    ports:
-      - "${PORT}:${PORT}" 
-
