+++ docker-compose.serve.yml
@@ -3,21 +3,21 @@ version: '3.8'
 services:
   ludwig:
     image: ludwigai/ludwig:master
-    command: >-
-      multi_serve_hf
-      --model_paths '${MODEL_PATHS}' --port ${PORT} --host ${HOST}
     stdin_open: true
     tty: true
-    env_file: .env
+    env_file: ./.env
+    command: >-
+      multi_serve
+      --model_paths '${MODEL_PATHS}' --mode ${MODE} --port ${PORT} --host ${HOST} 
     volumes:
       - ./data:/data 
       - ./results:/results
       - ./cli.py:/usr/local/lib/python3.8/site-packages/ludwig/cli.py
-      - ./multi_serve_hf.py:/usr/local/lib/python3.8/site-packages/ludwig/multi_serve_hf.py
-      - nlu-data:/app/repos
+      - ./multi_serve.py:/usr/local/lib/python3.8/site-packages/ludwig/multi_serve.py
+      - nlu-repos:/app/repos
 
     ports:
       - "${PORT}:${PORT}" 
 volumes:
-  nlu-data:
+  nlu-repos:
